requirements.txt
Flask
gunicorn
sklearn
pandas
numpy
pyYAML
-e .